CRM/requirements.txt
Django>=6.0.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-environ
psycopg2-binary
pandas
gunicorn
pytz